+++ packages/dart/lib/src/web.dart
@@ -1,3 +1,31 @@
+/// You can check if the URL path in the first argument value is matched against the second set of rules.
+bool isMatchPathname(String pathname, dynamic matcher) {
+  if (pathname.isEmpty ||
+      matcher == null ||
+      (matcher is String && matcher.isEmpty)) {
+    throw Exception('`url` and `matcher` must be set');
+  }
+
+  final List<String> matcherSet =
+      matcher is List<String> ? matcher : [matcher.toString()];
+  final String realPathname =
+      pathname.split('?').first.replaceFirst(RegExp(r'/$'), '');
+
+  return matcherSet.any((matcher) {
+    if (matcher == '*' || matcher == '/*') {
+      return true;
+    }
+
+    final realMatcher = matcher.replaceFirst(RegExp(r'\*$'), '');
+
+    if (matcher.endsWith('*')) {
+      return realPathname.startsWith(realMatcher);
+    }
+
+    return realMatcher == realPathname;
+  });
+}
+
 /// Removes the first-level path from a URL or pathname. Use this when you need a locale-free path in special cases in a URL that normally uses locale prefixes. For example, `/en/hello` is converted to `/hello`.
 String removeLocalePrefix(String urlOrPathname, dynamic locales) {
   final List<String> localeLists =