+++ packages/dart/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `headFile` method
 - Add `tailFile` method
 - Add `removeLocalePrefix` method
+- Add `isMatchPathname` method
 
 ## 1.1.8 (2025-11-13)